feat: implement server with background jobs and resume when all answers are provided (#49)
This implements #2 and #9 (please confirm)

Updated server to use latest API
Workflow now runs on the server in the background until it requests a
tool or finishes
Removed saving/loading from file as we want to keep state in memory and
run its updates in background. Previous example required calling
"iterate" endpoint each time to advance it. We can always add it back,
but I think this one is much better
Will automatically resume when all answers are provided 
Logger UI

![CleanShot 2024-12-09 at 04 26
13@2x](https://github.com/user-attachments/assets/8eaea18a-79c9-46a5-bc33-531be71c18b5)

Once this gets accepted, I will go ahead and create reusable functions
to reduce this boilerplate and hide abstraction.

---------

Co-authored-by: Piotr Karwatka <piotr.karwatka@gmail.com>