feat(create-fabrice-ai): downloading latest release (#116)
Before, we downloaded the `main` snapshot. It was fine. However, it
didn't work well when the current `main` source code was incompatible
with the released `npm` modules. This situation happened with the recent
#105. Now, we're getting the latest official release and taking the
examples out of it.

Because the tarball starts with a directory of an unknown name (in fact,
it's the `org-repo-lastcommitid` kind of) - we're about to extract the
full release and not only the example folder (as it was before) because
the leading directory name is unknown.

---------

Co-authored-by: Mike Grabowski <grabbou@gmail.com>